+++ packages/form-render/CHANGELOG.md
@@ -1,4 +1,15 @@
 # 更新日志
+### 2.0.13
+[+] 增加输入控件，extra、help 属性配置
+[+] 自定义校验 validator 支持返回一个 { status: true/ false , message: '错误信息' }，用于动态设置 message
+### 2.0.11
+[+] SearchForm 增加 layoutAuto 是否自适应布局
+### 2.0.10
+- [!] 修复 Modal 模式下，弹窗关闭，表单值无法清除
+
+### 2.0.9
+- [+] 优化 displayType inline 模式下表单项布局
+
 ### 2.0.6
 - [!] 修复 DatePicker 组件显示默认是英文提示
 ### 2.0.3