+++ packages/table-render/CHANGELOG.md
@@ -1,4 +1,11 @@
 # 更新日志
+### 2.0.9
+-[+] Search 增加极简模式 mode='simple'配置，直接通过onChange触发查询事件，并且去除查询按钮
+-[+] 增加 autoSize 是否自适应布局
+
+### 2.0.8
+-[+] valueType 新增 image、tag、tags、progress、dateRange、dateTimeRange 类型
+-[+] 支持列头配置气泡 tooltip 提示
 
 ### 2.0.5
 -[!] 修复 DatePicker 组件显示默认是英文提示